+++ internal/smart_compact/conversation_transform.go
