+++ internal/smart_compact/conversation_transform_test.go
