+++ internal/smart_compact/document_transform.go
