+++ internal/smart_compact/document_transform_test.go
