+++ internal/smart_compact/thinking_transform.go
@@ -7,14 +7,14 @@ import (
 	"github.com/tingly-dev/tingly-box/internal/protocol/transform"
 )
 
-// CompactTransform removes thinking fields from non-current rounds.
+// ThinkingCompactTransform removes thinking fields from non-current rounds.
 // This is the transform-style implementation that implements transform.Transform.
-type CompactTransform struct {
+type ThinkingCompactTransform struct {
 	rounder         *protocol.Grouper
 	keepLastNRounds int // Number of recent rounds to preserve thinking blocks (min: 1)
 }
 
-// NewCompactTransform creates a new CompactTransform.
+// NewCompactTransform creates a new ThinkingCompactTransform.
 //
 // The keepLastNRounds parameter controls how many recent conversation rounds
 // should have their thinking blocks preserved. Higher values retain more reasoning
@@ -28,19 +28,19 @@ func NewCompactTransform(keepLastNRounds int) transform.Transform {
 	if keepLastNRounds < 1 {
 		keepLastNRounds = 1
 	}
-	return &CompactTransform{
+	return &ThinkingCompactTransform{
 		rounder:         protocol.NewGrouper(),
 		keepLastNRounds: keepLastNRounds,
 	}
 }
 
 // Name returns the transform identifier.
-func (t *CompactTransform) Name() string {
+func (t *ThinkingCompactTransform) Name() string {
 	return "compact_thinking"
 }
 
 // Apply applies the compaction to the request.
-func (t *CompactTransform) Apply(ctx *transform.TransformContext) error {
+func (t *ThinkingCompactTransform) Apply(ctx *transform.TransformContext) error {
 	switch req := ctx.Request.(type) {
 	case *anthropic.MessageNewParams:
 		return t.applyV1(req)
@@ -53,7 +53,7 @@ func (t *CompactTransform) Apply(ctx *transform.TransformContext) error {
 }
 
 // applyV1 applies compaction to v1 requests.
-func (t *CompactTransform) applyV1(req *anthropic.MessageNewParams) error {
+func (t *ThinkingCompactTransform) applyV1(req *anthropic.MessageNewParams) error {
 	if req.Messages == nil || len(req.Messages) == 0 {
 		return nil
 	}
@@ -68,7 +68,7 @@ func (t *CompactTransform) applyV1(req *anthropic.MessageNewParams) error {
 }
 
 // applyBeta applies compaction to beta requests.
-func (t *CompactTransform) applyBeta(req *anthropic.BetaMessageNewParams) error {
+func (t *ThinkingCompactTransform) applyBeta(req *anthropic.BetaMessageNewParams) error {
 	if req.Messages == nil || len(req.Messages) == 0 {
 		return nil
 	}
@@ -99,7 +99,7 @@ func (t *CompactTransform) applyBeta(req *anthropic.BetaMessageNewParams) error 
 //	Rounds failing these checks are skipped (not compacted) to avoid corrupting
 //
 // potentially malformed conversation structures.
-func (t *CompactTransform) compactV1Rounds(rounds []protocol.V1Round) ([]anthropic.MessageParam, int) {
+func (t *ThinkingCompactTransform) compactV1Rounds(rounds []protocol.V1Round) ([]anthropic.MessageParam, int) {
 	var result []anthropic.MessageParam
 	removedCount := 0
 	totalRounds := len(rounds)
@@ -137,7 +137,7 @@ func (t *CompactTransform) compactV1Rounds(rounds []protocol.V1Round) ([]anthrop
 // compactBetaRounds removes thinking blocks from rounds outside the preservation window.
 //
 // See compactV1Rounds for detailed strategy rationale and guard checks.
-func (t *CompactTransform) compactBetaRounds(rounds []protocol.BetaRound) ([]anthropic.BetaMessageParam, int) {
+func (t *ThinkingCompactTransform) compactBetaRounds(rounds []protocol.BetaRound) ([]anthropic.BetaMessageParam, int) {
 	var result []anthropic.BetaMessageParam
 	removedCount := 0
 	totalRounds := len(rounds)
@@ -173,7 +173,7 @@ func (t *CompactTransform) compactBetaRounds(rounds []protocol.BetaRound) ([]ant
 }
 
 // removeV1ThinkingBlocks removes thinking content blocks from v1 message content.
-func (t *CompactTransform) removeV1ThinkingBlocks(content []anthropic.ContentBlockParamUnion, count int) ([]anthropic.ContentBlockParamUnion, int) {
+func (t *ThinkingCompactTransform) removeV1ThinkingBlocks(content []anthropic.ContentBlockParamUnion, count int) ([]anthropic.ContentBlockParamUnion, int) {
 	var filtered []anthropic.ContentBlockParamUnion
 
 	for _, block := range content {
@@ -189,7 +189,7 @@ func (t *CompactTransform) removeV1ThinkingBlocks(content []anthropic.ContentBlo
 }
 
 // removeBetaThinkingBlocks removes thinking content blocks from beta message content.
-func (t *CompactTransform) removeBetaThinkingBlocks(content []anthropic.BetaContentBlockParamUnion, count int) ([]anthropic.BetaContentBlockParamUnion, int) {
+func (t *ThinkingCompactTransform) removeBetaThinkingBlocks(content []anthropic.BetaContentBlockParamUnion, count int) ([]anthropic.BetaContentBlockParamUnion, int) {
 	var filtered []anthropic.BetaContentBlockParamUnion
 
 	for _, block := range content {
@@ -215,7 +215,7 @@ func (t *CompactTransform) removeBetaThinkingBlocks(content []anthropic.BetaCont
 //
 // Returns false if the round structure appears malformed, preventing compaction
 // on potentially incorrectly grouped rounds.
-func (t *CompactTransform) shouldCompactRound(stats *protocol.RoundStats) bool {
+func (t *ThinkingCompactTransform) shouldCompactRound(stats *protocol.RoundStats) bool {
 	// Guard: nil stats check
 	if stats == nil {
 		return false
